Fix noWrap mode on non-default aspect ratio screens

+++ app/src/main/java/com/betelge/rvlvr/gvr/GVRenderer.java
@@ -490,7 +490,7 @@ public class GVRenderer implements GvrView.StereoRenderer, DriftRenderer {
 
         if(noWrap && projectionType == DriftRenderer.PROJECTION_TYPE_NOVR) {
             Matrix.setIdentityM(mat, 0);
-            float yScale = height / viewHeight * viewWidth / width;
+            float yScale = height / (float)viewHeight * viewWidth / (float)width;
             Matrix.scaleM(mat, 0, noWrapZoom, noWrapZoom * yScale, 1);
             Matrix.translateM(mat, 0, noWrapX, noWrapY, 0);
             GLES20.glUniformMatrix4fv(mvpBlitLoc, 1, false, mat, 0);
